requirements.txt
appdirs==1.4.4
APScheduler==3.9.1
async-generator==1.10
attrs==21.4.0
bs4==0.0.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
cryptography==37.0.2
cssselect==1.1.0
fake-useragent==0.1.11
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
lxml==4.8.0
outcome==1.1.0
parse==1.19.0
pycparser==2.21
pyee==8.2.2
pyOpenSSL==22.0.0
pyppeteer==1.0.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyqt5-tools==5.15.4.3.2
pyquery==1.4.3
PySocks==1.7.1
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.27.1
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
w3lib==1.22.0
webdriver-manager==3.7.0
websockets==10.3
wsproto==1.1.0
zipp==3.8.0
